skycast_core_advance/requirements.txt
fastapi
uvicorn
sqlalchemy
requests
python-dotenv
fpdf2
httpx
chainlit
